app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
lark-oapi==1.4.0
protobuf==3.20.3
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.1
